+++ lib/src/mock_field_value_factory_platform.dart
@@ -1,4 +1,4 @@
-import 'package:cloud_firestore_mocks/src/mock_field_value_platform.dart';
+import 'package:fake_cloud_firestore/src/mock_field_value_platform.dart';
 import 'package:cloud_firestore_platform_interface/cloud_firestore_platform_interface.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';